refactor: 인플루언서 관리 조회 기능 수정
